pyproject.toml
[project]
name = "talk-to-data-slackbot"
version = "0.1.0"
description = "An AI agent for text queries on Slack."
authors = [
    {name = "Guilherme Bracco"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "litellm>=1.61.20,<2.0.0",
    "pandasai>=3.0.0",
    "pandasai-litellm",
    "pandasai-sql[postgres]",
    "python-dotenv>=1.0.0",
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
